fixed warning
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/aims/aims-free/branches/bug_fix@79893 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ cartodata/src/library/volume/volumebase.h
@@ -492,9 +492,9 @@ namespace carto
     Position4Di( const U & other ):
       _coords(4)
     {
-      for( int i = 0; i < other.size() && i < 4; ++i )
+      for( size_t i = 0; i < size_t( other.size() ) && i < 4; ++i )
         _coords[i] = other[i];
-      for( int i = other.size(); i < 4; ++i )
+      for( size_t i = size_t( other.size() ); i < 4; ++i )
         _coords[i] = 0;
     }
